utils - PR fixes - move relevant functions to reportUtil
